# docs/admin-api/admin_swagger.yaml
basePath: /api/v1
definitions:
  api.PageData:
    properties:
      list: {}
      pageNum:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
    type: object
  api.Response:
    properties:
      code:
        type: integer
      data: {}
      detail:
        type: string
      msg:
        type: string
      requestId:
        type: string
    type: object
...
  dto.SysPermissionTree:
    properties:
      anon:
        type: boolean
      auth:
        type: boolean
      children:
        items:
          $ref: '#/definitions/dto.SysPermissionTree'
        type: array
      createdAt:
        type: string
      createdBy:
        example: "0"
        type: string
      description:
        type: string
      id:
        example: "0"
        type: string
      key:
        example: "0"
        type: string
      name:
        type: string
      parentId:
        type: integer
      title:
        type: string
      type:
        type: string
      updatedAt:
        type: string
      updatedBy:
        example: "0"
        type: string
    type: object
...
  dto.SysRoleAddReq:
    properties:
      description:
        type: string
      name:
        type: string
    type: object
  dto.SysRoleBindPermissionsReq:
    properties:
      permissionIds:
        items:
          type: integer
        type: array
      roleId:
        example: "0"
        type: string
    type: object
  dto.SysRoleDeleteReq:
    properties:
      id:
        example: "0"
        type: string
    type: object
  dto.SysRoleGetByNameReq:
    properties:
      name:
        type: string
    type: object
  dto.SysRoleGetPermissionTreeReq:
    properties:
      roleId:
        example: "0"
        type: string
    type: object
  dto.SysRolePageQueryReq:
    properties:
      idOrder:
        type: string
      name:
        type: string
      pageNum:
        type: integer
      pageSize:
        type: integer
    type: object
  dto.SysRolePermissionTreeResp:
    properties:
      ownPermissionIds:
        items:
          type: integer
        type: array
      permissionTree:
        items:
          $ref: '#/definitions/dto.SysPermissionTree'
        type: array
    type: object
  dto.SysRoleUnbindUsersReq:
    properties:
      roleId:
        example: "0"
        type: string
      userIds:
        items:
          type: integer
        type: array
    type: object
  dto.SysRoleUpdateReq:
    properties:
      description:
        type: string
      id:
        example: "0"
        type: string
    type: object
  dto.SysSmsLogPageQueryReq:
    properties:
      createdAtEnd:
        type: string
      createdAtStart:
        type: string
      idOrder:
        type: string
      pageNum:
        type: integer
      pageSize:
        type: integer
      phone:
        type: string
      templateId:
        type: string
    type: object
  dto.SysSmsSendReq:
    properties:
      phone:
        type: string
      scene:
        type: string
    type: object
  dto.SysSmsSendToMyselfReq:
    properties:
      scene:
        type: string
    type: object
...
  models.SysRole:
    properties:
      createdAt:
        type: string
      createdBy:
        example: "0"
        type: string
      description:
        type: string
      id:
        example: "0"
        type: string
      name:
        type: string
      updatedAt:
        type: string
      updatedBy:
        example: "0"
        type: string
    type: object
  models.SysSmsLog:
    properties:
      content:
        type: string
      createdAt:
        type: string
      createdBy:
        example: "0"
        type: string
      id:
        example: "0"
        type: string
      phone:
        type: string
      remark:
        type: string
      scene:
        type: string
      status:
        type: string
      templateCode:
        type: string
      vendor:
        type: string
    type: object
...
host: 127.0.0.1:10000
info:
  contact:
    email: xxx@qq.com
    name: xxx
    url: http://xxx.com
  description: xxx管理后台API文档
  title: Admin API
  version: 1.0.0
paths:
...
  /sys/role/add:
    post:
      consumes:
      - application/json
      description: 添加角色
      parameters:
      - description: 角色信息
        in: body
        name: param
        schema:
          $ref: '#/definitions/dto.SysRoleAddReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - RequireLogin: []
      summary: 添加角色
      tags:
      - 角色管理
  /sys/role/bind_permissions:
    post:
      consumes:
      - application/json
      description: 批量关联权限给角色，会覆盖之前角色拥有的权限
      parameters:
      - description: 角色id和权限id数组
        in: body
        name: param
        required: true
        schema:
          $ref: '#/definitions/dto.SysRoleBindPermissionsReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - RequireLogin: []
      summary: 给角色授予权限
      tags:
      - 角色管理
  /sys/role/bind_users:
    post:
      consumes:
      - application/json
      description: 角色关联用户，关联后用户拥有该角色和该角色的所有权限
      parameters:
      - description: 需要绑定的角色id和用户id数组
        in: body
        name: param
        required: true
        schema:
          $ref: '#/definitions/dto.SysRoleBindPermissionsReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - RequireLogin: []
      summary: 授权角色给用户
      tags:
      - 角色管理
  /sys/role/delete:
    post:
      consumes:
      - application/json
      description: 根据id删除角色
      parameters:
      - description: 要删除的角色ID
        in: body
        name: param
        schema:
          $ref: '#/definitions/dto.SysRoleDeleteReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - RequireLogin: []
      summary: 删除角色
      tags:
      - 角色管理
  /sys/role/get_by_name:
    post:
      consumes:
      - application/json
      description: 根据名字获取角色
      parameters:
      - description: 角色名称
        in: body
        name: param
        schema:
          $ref: '#/definitions/dto.SysRoleGetByNameReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - RequireLogin: []
      summary: 根据名字获取角色
      tags:
      - 角色管理
  /sys/role/get_permission_tree:
    post:
      consumes:
      - application/json
      description: 获取角色拥有的权限树
      parameters:
      - description: 角色id
        in: body
        name: param
        required: true
        schema:
          $ref: '#/definitions/dto.SysRoleGetPermissionTreeReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.SysRolePermissionTreeResp'
              type: object
      security:
      - RequireLogin: []
      summary: 获取角色拥有的权限树
      tags:
      - 角色管理
  /sys/role/page_query:
    post:
      consumes:
      - application/json
      description: 分页查询角色
      parameters:
      - description: 角色筛选条件
        in: body
        name: param
        schema:
          $ref: '#/definitions/dto.SysRolePageQueryReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/api.PageData'
                  - properties:
                      List:
                        $ref: '#/definitions/models.SysRole'
                    type: object
              type: object
      security:
      - RequireLogin: []
      summary: 分页查询角色
      tags:
      - 角色管理
  /sys/role/unbind_users:
    post:
      consumes:
      - application/json
      description: 取消角色关联用户，取消后用户不再拥有该角色
      parameters:
      - description: 需要解绑的角色id和用户id数组
        in: body
        name: param
        required: true
        schema:
          $ref: '#/definitions/dto.SysRoleUnbindUsersReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - RequireLogin: []
      summary: 取消授权角色给用户
      tags:
      - 角色管理
  /sys/role/update:
    post:
      consumes:
      - application/json
      description: 根据id修改角色信息
      parameters:
      - description: 角色信息
        in: body
        name: param
        schema:
          $ref: '#/definitions/dto.SysRoleUpdateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - RequireLogin: []
      summary: 修改角色
      tags:
      - 角色管理
  /sys/sms/page_query:
    post:
      consumes:
      - application/json
      description: 分页查询短信记录
      parameters:
      - description: 短信记录筛选条件
        in: body
        name: param
        schema:
          $ref: '#/definitions/dto.SysSmsLogPageQueryReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/api.PageData'
                  - properties:
                      List:
                        $ref: '#/definitions/models.SysSmsLog'
                    type: object
              type: object
      security:
      - RequireLogin: []
      summary: 分页查询短信记录
      tags:
      - 短信管理
  /sys/sms/send:
    post:
      consumes:
      - application/json
      description: 发送短信
      parameters:
      - description: 手机号和发送场景
        in: body
        name: param
        schema:
          $ref: '#/definitions/dto.SysSmsSendReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response'
      summary: 发送短信
      tags:
      - 短信管理
  /sys/sms/send_to_myself:
    post:
      consumes:
      - application/json
      description: 发送短信(登录后)
      parameters:
      - description: 发送场景
        in: body
        name: param
        schema:
          $ref: '#/definitions/dto.SysSmsSendToMyselfReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - RequireLogin: []
      summary: 发送短信(登录后)
      tags:
      - 短信管理
...
securityDefinitions:
  RequireLogin:
    in: header
    name: token
    type: apiKey
swagger: "2.0"